# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple API to demonstrate Swagger integration in Go with Fiber.",
        "title": "My API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/public": {
            "get": {
                "description": "This is a public route that is accessible by everyone",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Public route",
                "responses": {
                    "200": {
                        "description": "Hello, Public Page....!!!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new public resource in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Create public resource",
                "parameters": [
                    {
                        "description": "Resource creation data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CreatePublicRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Resource created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Uploads a file to the server",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "File Upload"
                ],
                "summary": "Upload a file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Upload successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Failed to upload file",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "main.CreatePublicRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Resource description"
                },
                "name": {
                    "type": "string",
                    "example": "Resource name"
                }
            }
        },
        "main.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "Error message"
                }
            }
        }
    }
}